api model update
