+++ Pit Hub/Services/GPManager.swift
@@ -24,7 +24,7 @@ struct GPManager {
         
         do {
             let scheduleResponse = try decoder.decode(F1ScheduleResponse.self, from: data)
-            let races = scheduleResponse.mrData.raceTable.races
+            let races = scheduleResponse.mrData.raceTable?.races ?? []
             print("Successfully fetched \(races.count) races for \(year). ------------- ")
             return races
         } catch {
